Matrix buffer data vector changed to float to avoid type casting in CUDA code. BLAS operations can use only PCudaMatrixBuffer now (need a better solution...)

+++ c_src/pcuda_buffer.h
@@ -152,9 +152,9 @@ public:
     virtual void write(ErlNifEnv *env, ERL_NIF_TERM data);
     virtual void clear();
     
-    std::vector<double>* get_data(){return data;};
+    std::vector<float>* get_data(){return data;};
 protected:
-    std::vector<double> *data;
+    std::vector<float> *data;
 };
 
 #endif
